+++ NEWS.md
@@ -1,8 +1,20 @@
 # odbc (development version)
 
+* odbcConnectionColumns(), odbcConnectionIcon(), and odbcConnectionActions() are now fully deprecated.
+  Use DBI::dbListFields() instead of odbcConnectionColumns() (#699).
+# odbc 1.6.4
+
+* Fix writing of [R] date/time values that have integer storage. (#952)
+
+* Oracle: Fix writing values to `DATE` targets. (#959)
+
+* SQL Server: Fix time zone interpretation in `DATETIMEOFFSET` data; now follows
+  ISO 8061 convention where positive offset denotes time zone east of 
+  Greenwich. (#946)
+
 # odbc 1.6.3
 
-Addressed a compiler warning on `r-devel-linux-x86_64-fedora-clang` (#941).
+* Addressed a compiler warning on `r-devel-linux-x86_64-fedora-clang` (#941).
 
 # odbc 1.6.2